Add --custom-folder
Add --custom-folder and folder source to allow creating shortcuts from
files in a single folder. Useful for files launched by a single program
(interpreted games, video files, etc.) or for a bunch of executables in
a folder.

The folder source doesn't really make sense on its own since it needs
the folder specified, but we select all of our actions based on sources
and default to all of them so we need a way for users to select only the
folder source.