Allow 'quit' in game selection prompt
I often change my mind at this point and the only way to get out is to
Ctrl-c. Add a quit command so you can type 'q' or 'quit' to cancel
selection and abort.

Alternatives: Could return "quit" or [] or adding a second return value,
but returning False seems cleaner.